lb_pair
Module execute

Structs

MintArrays

Statics

MAX_TOTAL_FEE 🔒

Functions

_get_flash_loan_fees

Returns the encoded fees amounts for a flash loan

approx_div

batch_transfer_from

Overrides the batch transfer function to call the hooks before and after the transfer

burn

Burn Liquidity Book (LB) tokens and withdraw tokens from the pool.

collect_protocol_fees

Collect the protocol fees from the pool.

flash_loan

Flash loan tokens from the pool to a receiver contract and execute a callback function. The receiver contract is expected to return the tokens plus a fee to this contract. The fee is calculated as a percentage of the amount borrowed, and is the same for both tokens.

force_decay

Forces the decay of the volatility reference variables.

increase_oracle_length

Increase the length of the oracle used by the pool.

mint

Mint liquidity tokens by depositing tokens into the pool.

mint_bins 🔒

Helper function to mint liquidity in each bin in the liquidity configurations.

set_hooks_parameters

@notice Sets the hooks parameter of the pool @dev Can only be called by the factory @param hooksParameters The hooks parameter @param onHooksSetData The data to be passed to the onHooksSet function of the hooks contract

set_static_fee_parameters

Sets the static fee parameters of the pool.

swap

Swap tokens iterating over the bins until the entire amount is swapped.

update_bin 🔒

Helper function to update a bin during minting.