lb_pair::execute
Function flash_loan

pub fn flash_loan(
    deps: DepsMut<'_>,
    _env: Env,
    info: MessageInfo,
    receiver: ContractInfo,
    amounts: Bytes32,
    data: Option<Binary>,
) -> Result<Response>

Flash loan tokens from the pool to a receiver contract and execute a callback function. The receiver contract is expected to return the tokens plus a fee to this contract. The fee is calculated as a percentage of the amount borrowed, and is the same for both tokens.

Arguments

receiver - The contract that will receive the tokens and execute the callback function

amounts - The encoded amounts of token X and token Y to flash loan

data - Any data that will be passed to the callback function