lb_pair::execute
Function mint

pub fn mint(
    deps: DepsMut<'_>,
    env: Env,
    info: MessageInfo,
    to: String,
    liquidity_configs: Vec<LiquidityConfigurations>,
    refund_to: String,
) -> Result<Response>

Mint liquidity tokens by depositing tokens into the pool.

It will mint Liquidity Book (LB) tokens for each bin where the user adds liquidity. This function will not transfer the tokens from the caller, it is expected that the tokens have already been transferred to this contract through another contract, most likely the router. That is why this function shouldn’t be called directly, but through one of the add liquidity functions of a router that will also perform safety checks.

Any excess amount of token will be sent to the refund_to address.

Arguments

to - The address that will receive the LB tokens

liquidity_configs - The encoded liquidity configurations, each one containing the id of the bin and the percentage of token X and token Y to add to the bin.

refund_to - The address that will receive the excess amount of tokens

Returns

amounts_received - The amounts of token X and token Y received by the pool

amounts_left - The amounts of token X and token Y that were not added to the pool and were sent to to

liquidity_minted - The amounts of LB tokens minted for each bin