lb_pair::execute
Function swap

pub fn swap(
    deps: DepsMut<'_>,
    env: Env,
    info: MessageInfo,
    swap_for_y: bool,
    to: String,
) -> Result<Response>

Swap tokens iterating over the bins until the entire amount is swapped.

Token X will be swapped for token Y if swap_for_y is true, and token Y for token X if swap_for_y is false.

This function will not transfer the tokens from the caller, it is expected that the tokens have already been transferred to this contract through another contract, most likely the router. That is why this function shouldn’t be called directly, but only through one of the swap functions of a router that will also perform safety checks, such as minimum amounts and slippage.

The variable fee is updated throughout the swap, it increases with the number of bins crossed. The oracle is updated at the end of the swap.

Arguments

swap_for_y - Whether you’re swapping token X for token Y (true) or token Y for token X (false)

to - The address to send the tokens to

Returns

amounts_out - The encoded amounts of token X and token Y sent to to