lb_pair
Module helper

Functions

_balance_of

_get_next_non_empty_bin

Returns id of the next non-empty bin.

_get_total_supply

bin_transfer

Transfers the encoded amounts to the recipient for both tokens.

bin_transfer_x

Transfers the encoded amounts to the recipient, only for token X.

bin_transfer_y

Transfers the encoded amounts to the recipient, only for token Y.

only_factory

only_protocol_fee_recipient

query_token_symbol

register_pair_token