lb_pair
Module lb_token

Enums

Error

Statics

BALANCES 🔒

The mapping from account to token id to account balance.

SPENDER_APPROVALS 🔒

Mapping from account to spender approvals.

TOTAL_SUPPLIES 🔒

The mapping from token id to total supply.

TOTAL_SUPPLY 🔒

global total supply tracker

Functions

_approve_for_all 🔒

_batch_transfer_from 🔒

_burn 🔒

Burn amount of id from account. The account must not be the zero address. The event should be emitted by the contract that inherits this contract.

_check_length 🔒

_get_all_user_bins 🔒

_is_approved_for_all 🔒

Returns true if spender is approved to transfer owner’s tokens or if spender is the owner.

_mint 🔒

Mint amount of id to account. The account must not be the zero address. The event should be emitted by the contract that inherits this contract.

_not_address_zero_or_this 🔒

approve_for_all

Grants or revokes permission to spender to transfer the caller’s tokens, according to approved.

balance_of

Returns the amount of tokens of type id owned by account.

balance_of_batch

Return the balance of multiple (account/id) pairs.

batch_transfer_from

Batch transfers amounts of ids from from to to.

check_approval

Modifier to check if the spender is approved for all.

check_length

Modifier to check if the length of the arrays are equal.

is_approved_for_all

Returns true if spender is approved to transfer owner’s tokens or if spender is the owner.

name

Returns the name of the token.

not_address_zero_or_this

Modifier to check if the address is not zero or the contract itself.

symbol

Returns the symbol of the token, usually a shorter version of the name.

total_supply

Returns the total supply of token of type id.

Type Aliases

Result

Alias for Result<T, LbTokenError>