lb_pair
Module query

Functions

get_active_id

Returns the active id of the Liquidity Book Pair.

get_all_bins

Returns the reserves of all bins, paginated.

get_bin

Returns the reserves of a bin.

get_bin_step

Returns the bin step of the Liquidity Book Pair.

get_bins

Returns the reserves of many bins.

get_factory

Returns the Liquidity Book Factory.

get_id_from_price

Returns the id corresponding to the given price.

get_lb_hooks_parameters

Gets the hooks parameters of the Liquidity Book Pair.

get_lb_token

Returns the LB Token of the Liquidity Book Pair.

get_next_non_empty_bin

Returns the next non-empty bin.

get_oracle_params

Returns the oracle parameters of the Liquidity Book Pair.

get_oracle_sample_at

Returns the cumulative values of the Liquidity Book Pair at a given timestamp.

get_price_from_id

Returns the price corresponding to the given id, as a 128.128-binary fixed-point number.

get_protocol_fees

Returns the protocol fees of the Liquidity Book Pair.

get_reserves

Returns the reserves of the Liquidity Book Pair.

get_static_fee_parameters

Returns the static fee parameters of the Liquidity Book Pair.

get_swap_in

Simulates a swap in.

get_swap_out

Simulates a swap out.

get_token_x

Returns the token X of the Liquidity Book Pair.

get_token_y

Returns the token Y of the Liquidity Book Pair.

get_total_supply

Returns the total supply of the associated LB Token.

get_variable_fee_parameters

Returns the variable fee parameters of the Liquidity Book Pair.