lb_pair::query
Function get_swap_in

pub fn get_swap_in(
    deps: Deps<'_>,
    env: Env,
    amount_out: u128,
    swap_for_y: bool,
) -> Result<SwapInResponse>

Simulates a swap in.

Note

If amount_out_left is greater than zero, the swap in is not possible, and the maximum amount that can be swapped from amountIn is amountOut - amountOutLeft.

Arguments

amount_out - The amount of token X or Y to swap in

swap_for_y - Whether the swap is for token Y (true) or token X (false)

Returns

amount_in - The amount of token X or Y that can be swapped in, including the fee

amount_out_left - The amount of token Y or X that cannot be swapped out

fee - The fee of the swap